src: Rename HelperFunctions_DAC.h to HelperFunctions_ITC.h
